moved swapElements to utils
